requirements.txt
python-dotenv==1.0.0
supabase==2.0.3
rich==13.7.0
tabulate==0.9.0
postgrest==0.13.0